+++ src/openai/resources/audio/transcriptions.py
@@ -143,7 +143,7 @@ class Transcriptions(SyncAPIResource):
 
           prompt: An optional text to guide the model's style or continue a previous audio
               segment. The
-              [prompt](https://platform.openai.com/docs/guides/speech-to-text/prompting)
+              [prompt](https://platform.openai.com/docs/guides/speech-to-text#prompting)
               should match the audio language.
 
           response_format: The format of the output, in one of these options: `json`, `text`, `srt`,
@@ -307,7 +307,7 @@ class AsyncTranscriptions(AsyncAPIResource):
 
           prompt: An optional text to guide the model's style or continue a previous audio
               segment. The
-              [prompt](https://platform.openai.com/docs/guides/speech-to-text/prompting)
+              [prompt](https://platform.openai.com/docs/guides/speech-to-text#prompting)
               should match the audio language.
 
           response_format: The format of the output, in one of these options: `json`, `text`, `srt`,