+++ src/openai/resources/moderations.py
@@ -68,7 +68,7 @@ class Moderations(SyncAPIResource):
           model: The content moderation model you would like to use. Learn more in
               [the moderation guide](https://platform.openai.com/docs/guides/moderation), and
               learn about available models
-              [here](https://platform.openai.com/docs/models/moderation).
+              [here](https://platform.openai.com/docs/models#moderation).
 
           extra_headers: Send extra headers
 
@@ -138,7 +138,7 @@ class AsyncModerations(AsyncAPIResource):
           model: The content moderation model you would like to use. Learn more in
               [the moderation guide](https://platform.openai.com/docs/guides/moderation), and
               learn about available models
-              [here](https://platform.openai.com/docs/models/moderation).
+              [here](https://platform.openai.com/docs/models#moderation).
 
           extra_headers: Send extra headers